move local / autograd to ComponentModeler.run()
